organize client files
