+++ src/App.jsx
@@ -1,10 +1,10 @@
-import axios from 'axios';
-
 import React, { Component } from 'react';
-import { BrowserRouter as Router, Route, Switch, Link } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 
 import Navbar from './navbar/Navbar';
-import Login, { setAuthToken } from './login/Login';
+import Landing from './landing/Landing';
+import Login from './login/Login';
+import Profile from './profile/Profile';
 
 class App extends Component {
   render() {
@@ -23,69 +23,6 @@ class App extends Component {
           </div>
         </div>
       </Router>
-    );
-  }
-}
-
-class Landing extends Component {
-  render() {
-    return (
-      <div className="tagline">
-        <div>
-          What are you <b>reading</b>?
-        </div>
-        <div>
-          Let's
-          <Link to="/login" className="link">
-            start a diary
-          </Link>
-          and show your friends.
-        </div>
-      </div>
-    );
-  }
-}
-
-class Profile extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { books: [], isLoaded: false };
-  }
-
-  componentDidMount() {
-    setAuthToken();
-    axios
-      .get('/user/profile')
-      .then(res => {
-        const user = res.data;
-        this.setState({ isLoaded: true });
-
-        axios
-          .get('/api/books', { params: { bookIds: user.bookIds } })
-          .then(res => {
-            this.setState({ books: res.data });
-          })
-          .catch();
-      })
-      .catch();
-  }
-
-  render() {
-    if (!this.state.isLoaded) {
-      return <div className="book-shelf"> Loading... </div>;
-    }
-
-    return (
-      <div>
-        <div className="book-shelf">Your Bookshelf</div>
-        <ul>
-          {this.state.books.map((book, index) => (
-            <li key={index} className="book-container">
-              {book.title}
-            </li>
-          ))}
-        </ul>
-      </div>
     );
   }
 }
