+++ src/index.scss
@@ -1,21 +1,4 @@
-$dark: #2c3340;
-$blue: #6b9f9f;
-$cream: #e5d3b5;
-$white: #ffffff;
-
-$serif: Georgia;
-$sans-serif: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen-Sans,
-  Ubuntu, Cantarell, 'Helvetica Neue', sans-serif;
-
-$page-width: 1000px;
-
-$text-7: 35px;
-$text-6: 29px;
-$text-5: 24px;
-$text-4: 20px;
-$text-3: 16px;
-$text-2: 14px;
-$text-1: 12px;
+@import 'template.scss';
 
 html,
 body,
@@ -214,49 +197,6 @@ button {
     background-color: rgba($dark, 0.9);
     box-shadow: 0 7px 14px rgba($blue, 0.7), 0 3px 6px rgba($blue, 0.9);
     transform: translateY(-1px);
-  }
-}
-
-.book-shelf {
-  text-align: center;
-  font-size: 4.5em;
-  font-weight: 600;
-  margin-bottom: 0.6em;
-  margin-top: 0.3em;
-  color: $dark;
-  text-shadow: 0 7px 14px rgba($blue, 0.2), 0 3px 6px rgba($blue, 0.3);
-}
-
-.book-container {
-  font-size: 1.3em;
-  font-weight: 600;
-  margin-bottom: 1em;
-  margin: 1em auto;
-  padding: 0.8em 1.2em;
-  background-color: $dark;
-  color: $cream;
-  border-radius: 4px;
-  width: 50%;
-  box-shadow: 0 7px 14px rgba($blue, 0.2), 0 3px 6px rgba($blue, 0.3);
-}
-
-.tagline {
-  text-align: center;
-  font-size: 3em;
-  margin-top: 18vh;
-  line-height: 2;
-
-  b {
-    font-weight: 600;
-  }
-
-  .link {
-    font-weight: 600;
-    box-shadow: 0 7px 14px rgba($blue, 0.2), 0 3px 6px rgba($blue, 0.3);
-    background-color: $dark;
-    padding: 5px 15px;
-    color: $cream;
-    margin: 0 10px;
   }
 }
