+++ src/login/Login.scss
@@ -1,4 +1,4 @@
-@import 'index.scss';
+@import 'template.scss';
 
 .login-form {
   width: 400px;
