+++ src/navbar/Navbar.scss
@@ -1,4 +1,4 @@
-@import 'index.scss';
+@import 'template.scss';
 
 .navbar-container {
   background-color: rgba($white, 0.1);
