+++ src/profile/Profile.scss
@@ -1,0 +1,24 @@
+@import 'template.scss';
+
+.book-shelf {
+  text-align: center;
+  font-size: 4.5em;
+  font-weight: 600;
+  margin-bottom: 0.6em;
+  margin-top: 0.3em;
+  color: $dark;
+  text-shadow: 0 7px 14px rgba($blue, 0.2), 0 3px 6px rgba($blue, 0.3);
+}
+
+.book-container {
+  font-size: 1.3em;
+  font-weight: 600;
+  margin-bottom: 1em;
+  margin: 1em auto;
+  padding: 0.8em 1.2em;
+  background-color: $dark;
+  color: $cream;
+  border-radius: 4px;
+  width: 50%;
+  box-shadow: 0 7px 14px rgba($blue, 0.2), 0 3px 6px rgba($blue, 0.3);
+}
