+++ src/template.scss
@@ -1,0 +1,18 @@
+$dark: #2c3340;
+$blue: #6b9f9f;
+$cream: #e5d3b5;
+$white: #ffffff;
+
+$serif: Georgia;
+$sans-serif: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen-Sans,
+  Ubuntu, Cantarell, 'Helvetica Neue', sans-serif;
+
+$page-width: 1000px;
+
+$text-7: 35px;
+$text-6: 29px;
+$text-5: 24px;
+$text-4: 20px;
+$text-3: 16px;
+$text-2: 14px;
+$text-1: 12px;
